nn: remove the obsolete CRNN model
The model was an experiment to explore whether we can build a network
that is superior to what we found in the papers. It worked, but is
was not better in any way than the CNN or CRNN. Moreover it relied on
a custom hack on TensorFlow, which is not present in the library nowadays,
so it's actually impossible to run. Removing it, will save some confusion.