nn: substitute accuracy metric with top_k_accuracy
Accuracy metric is bad for problems, where the target vector is really large
e.g. if we predict from 10k keywords and usually 1-5 are correct,
then absolute accuracy will always be in the 99,9...% range even if the model
predicts only 0s.

Top K accuracy looks at the top K (K=5 in this case I think) predictions
and checks whether the correct answer is in this set. This reflects better
whether a model performs well or not.